backend/requirements.txt
langchain==0.3.7
langchain-community==0.3.7
langchain-mistralai==0.2.12
mistralai==1.1.0
# langsmith is installed as a dependency of langchain, no need to specify separately

# Vector Store and Embeddings
chromadb==0.5.11
sentence-transformers==3.2.1
huggingface_hub==0.26.1

# PDF Processing
pypdf==4.3.1

# HTTP Requests
requests==2.32.3
httpx==0.27.2
aiohttp==3.10.5
tenacity==8.3.0

# Utilities
python-dotenv==1.0.1
tiktoken==0.7.0
rapidfuzz==3.9.7
ujson==5.9.0
loguru==0.7.2

# FastAPI (required for the web server)
fastapi==0.115.0
uvicorn[standard]==0.30.6
starlette==0.38.6
python-multipart==0.0.9
pydantic==2.9.2
